Add a function to retrive the merge base of the Pull Request.
The problem with GitHub's 'base' in the Pull Request webhook payload or
API call is that it tracks the commit in the "main" branch at which the
Pull Request was opened, not the commit that was used to actually create
the branch that is used to send the Pull Request.

In order to fix this problem and the inconsistencies it introduces, this
new function finds the first common ancestor (i.e. merge base) using
GitHub APIs. If this turns to be too slow we can of course switch to
using a local tree.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>