action.py: adjust deprecated backslash usage
The re_url regular expression string is valid because '\'
followed by something that is not an re special character, ASCII
letter, or ASCII number matches the following character.

Therefore, '\/' as a regular expression just matches the character
'/'.

But this usage is now deprecated because it's hard to read, and will
cause a syntax error  in the future:

   Also, please note that any invalid escape sequences in Python’s
   usage of the backslash in string literals now generate a
   DeprecationWarning and in the future this will become a
   SyntaxError.

   https://docs.python.org/3/library/re.html

Clean it up by removing the deprecated backslashes in one case that's
not a raw string, which is causing a linter error, and in another that
is, just for consistency.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>