action: Fix accessing manifest files in private repos
Whenever the manifest action is used in a private repository, then
authorization is required in order to obtain the contents of the two
revisions of the manifest file (old and new).
In order for authorization to work properly, one needs a download URL in the
style of:

https://raw.githubusercontent.com/<org>/<repo>/<sha>/<file>?token=<TOKEN>

Raw URLs like these will *not* work with private repos:

https://github.com/<org>/<repo>/raw/<sha>/<file>

where TOKEN is a temporary token generated by GitHub itself (this is NOT
the usual PAT).

For the old file there's no issue since Repository.get_contents()
already returns a download URL in the format above.
For the new file however, the raw_url that we were using so far will not
work. Instead, use the contents_url to retrieve a ContentFile that then
contains a download_url.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>