manifest: Add support for manifest comparison using tree checkouts
The current mechanism to compare manifests using manifest files directly
downloaded from GitHub works well when the main manifest does not import
any other local manifest. But when the self: section of the manifest
imports additional local manifests then a full tree checkout is required
for West to fully resolve the manifest.

Add a new use-tree-checkout parameter that forces the script to
explicitly check out a revision of the manifest repository to construct
the Manifest object in order to support this use case.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>