action: Change the logic to handle DNM labels
Bail out early if the comment variable is not set, since that means that
there is a guarantee nothing needs to be done at all. Then simplify the
logic to handle the removal of DNM labels back to its original form: if
there are any added or PR-pointing projects then add the DNM, otherwise
remove it.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>