action: Add a new option to check for impostor commits
On GitHub, commits can appear to be in a repo but in fact only be
present in a fork. To avoid letting users point to these type of
commits, named impostor commits, add an optional check that will not
remove the DNM label if one is present.

More info here:
https://www.chainguard.dev/unchained/what-the-fork-imposter-commits-in-github-actions-and-ci-cd

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>