action: Check additional metadata in projects
Until now this action was only looking at differences in revision for
all the projects in the manifest. However, there are other items of
metadata that can be highly sensitive, and changes to those must be
monitored and reported.

This commit introduces additional checks for the following metadata
items:

- URL: part of the manifest, URL to the project repository
- Submodules: part of the manifest, enabled submodules when running west
  update
- West commands: part of the manifest, pointer to a file defining
  extensions commands for west itself
- module.yml: the Zephyr module metadata file

If any of the items above is added, removed or changed, the DNM label
will not be removed from the Pull Request, in order for a human to
decide if the PR is to be merged or not.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>