action: Allow the user to set the west manifest import flag
In some cases, when using checked out trees to construct the manifest
objects, it is important for users to be able to select which projects
are imported, because they may have import statements in their manifests
that they may want to honour or not.
This commit allows the action user to select between
import all projects (the default), none or only the projects in the
self: section of the manifest.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>