action: Handle parsing of module.yml and binary blobs
This commit adds support for handling the differences inside the
module.yml file in a west project update, as well as locating and
displaying any changes to binary blobs.
More specifically, the metadata table now displays the number of added
and/or changed blobs per project, and optionall labels can be added to
the Pull Request to signal reviewers that blobs have been altered.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>