action: Fix blob sets so they use the path as an identifier
When converting blob dictionaries to tuples in order to be able to use sets,
until now we were sorting by key order, meaning typically "description"
was the item that ended up in the first place of each tuple in the set.
This is an issue because the _get_sets() function uses the first item of
each tuple in the set as the identifier of an item to detect duplicates.
Instead now, when populating the blob sets ensure that "path"
is the first in each tuple (of tuples) that is then converted to a set.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>