action: Check out the PR head in the Python implementation
Checking out the PR HEAD SHA in the workflows using this action, like so:

ref: ${{ github.event.pull_request.head.sha }}

is an issue when using the pull_request_target trigger. This is because
the workflows in this case run with write permissions to the repo, and
blindly checking out the PR contents instead of the base branch is
dangerous.

In order to avoid this issue, the Python code now fetches (and
temporarily checks out) the PR itself, so that the workflow invoking it
can just check out the regular base branch instead.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>